Bug when status 'incomplete' but score 'passes'
Test added and base logic changed. Even if the score is submitted, and
it also exceeds the mastery_score, the status may only be set to
'passed'/'complete'/'failed' if the course (or other logic) has
indicated that the course has actually been completed by the student.
For HACP, the API will be supplied a value of 'incomplete'. For the JS
API, the server side will receive a pre-computed lesson_status that does
not require server logic.